--- examples/TestBlazorWebAppNet8/Components/Pages/Teachers/Teacher.razor
@@ -1,58 +1,0 @@
-@page "/teachers"
-@using DataTransferObjects.Teachers
-@using Services.Teachers
-@inject ITeachersService TeacherService
-
-<h3>Teacher List</h3>
-
-<NavLink class="btn btn-primary mb-3" href="/teachers/create">Create New</NavLink>
-
-@if (teachers == null)
-{
-    <p><em>Loading...</em></p>
-}
-else
-{
-    <table class="table">
-        <thead>
-            <tr>
-                <th>UserId</th>
-                <th>User</th>
-                <th>Students</th>
-                <th>Id</th>
-                <th>Actions</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var item in teachers)
-            {
-                <tr>
-                    <td>@item.UserId</td>
-                    <td>@item.User</td>
-                    <td>@item.Students</td>
-                    <td>@item.Id</td>
-                    <td>
-                        <NavLink class="btn btn-sm btn-info" href="@($"/teachers/details/{item.Id}")">Details</NavLink>
-                        <NavLink class="btn btn-sm btn-warning" href="@($"/teachers/edit/{item.Id}")">Edit</NavLink>
-                        <button class="btn btn-sm btn-danger" @onclick="() => DeleteTeacher(item.Id)">Delete</button>
-                    </td>
-                </tr>
-            }
-        </tbody>
-    </table>
-}
-
-@code {
-    private List<TeacherViewModel>? teachers;
-
-    protected override async Task OnInitializedAsync()
-    {
-        teachers = await TeacherService.GetAllAsync();
-    }
-
-    async Task DeleteTeacher(Int32 id)
-    {
-        await TeacherService.DeleteAsync(id);
-        teachers = await TeacherService.GetAllAsync(); // Refresh list
-    }
-}
