+++ examples/TestBlazorWebAppNet8/Components/Pages/Users/User.Details.razor
@@ -1,4 +1,4 @@
-@page "/users/details/{Id:int}"
+@page "/users/{Id:int}"
 @using DataTransferObjects.Users
 @using Services.Users
 @inject IUsersService UserService
@@ -6,42 +6,40 @@
 
 <h3>User Details</h3>
 
-@if (userDetails == null)
+@if (user == null)
 {
     <p><em>Loading...</em></p>
 }
 else
 {
     <dl class="row">
-        <dt class="col-sm-3">Id</dt>
-        <dd class="col-sm-9">@userDetails.Id</dd>
         <dt class="col-sm-3">Name</dt>
-        <dd class="col-sm-9">@userDetails.Name</dd>
+        <dd class="col-sm-9">@user.Name</dd>
         <dt class="col-sm-3">Email</dt>
-        <dd class="col-sm-9">@userDetails.Email</dd>
+        <dd class="col-sm-9">@user.Email</dd>
         <dt class="col-sm-3">Password</dt>
-        <dd class="col-sm-9">@userDetails.Password</dd>
+        <dd class="col-sm-9">@user.Password</dd>
     </dl>
     <div>
-        <button class="btn btn-primary" @onclick="EditUser">Edit</button>
+        <button class="btn btn-primary" @onclick="Save">Save</button>
         <button class="btn btn-secondary" @onclick="BackToList">Back to List</button>
     </div>
 }
 
 @code {
     [Parameter]
-    public Int32 Id { get; set; }
+    public int Id { get; set; }
 
-    private UserViewModel? userDetails;
+    private UserViewModel? user;
 
     protected override async Task OnParametersSetAsync()
     {
-        userDetails = await UserService.GetByIdAsync(Id);
+        user = await UserService.GetByIdAsync(Id);
     }
 
-    private void EditUser()
+    private void Save()
     {
-        NavigationManager.NavigateTo($"/users/edit/{Id}");
+        throw new NotImplementedException("Save functionality is not implemented.");
     }
 
     private void BackToList()
